assets/04_내부업무포털/src/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-auth-ldap==4.6.0
python-ldap==3.4.4
django-cors-headers==4.3.1
django-filter==23.5
python-dotenv==1.0.0
whitenoise==6.6.0
dj-database-url==2.1.0
